+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.5 (in developing)
 - Package more clean now
+- Changed namespace from `s1lver\summernote` to `S1lver\Summernote`
 
 ## 1.0.4
 - First release